requirements.txt
# Web server
flask
flask-cors
gunicorn

# Túnel HTTP
pyngrok

# WhatsApp via Twilio
twilio

# IA (opcional - PydanticAI)
pydantic-ai

# Banco de Dados MySQL
mysql-connector-python

# Manipulação de dados CSV
pandas
python-dateutil
unidecode

# Variáveis de ambiente (.env)
python-dotenv

# Integração direta com GPT
openai

# Requisições HTTP
requests

# Testes automatizados
pytest
